MNT: move ordereddefaultdict to utils
